Moves helper functions into lib file
